% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.matrix.R
\name{inverse.matrix}
\alias{inverse.matrix}
\alias{inverse.vector}
\title{Pseudo-Inverse of Numeric Vectors and Matrices}
\usage{
inverse.matrix (X, eps = NULL, ...)

inverse.vector (X, ...)
}
\arguments{
\item{X}{numeric vector or (non-square) matrix.}

\item{eps}{numeric, small positive real in the open \code{(0, 0.1)}; Or
\code{NULL} (see section Details). \code{eps} is a positive tolerance value
below which any singular value of a matrix is considered zero.}

\item{...}{additional argument passed to or from other methods. Currently not
used.}
}
\value{
a numeric \code{'matrix'} if \code{object} is of class
\code{matrix}, and a numeric \code{'vector'} if \code{object} is of class
\code{vector}.

The returned object also has attributes:
\itemize{
\item \code{'m'} which is the minimum of the numbers of rows (if any) and column (if any)
of \code{object} (\code{m = 1} for any vector);
\item \code{'rank'} which is the matrix rank of \code{object} or its returned inverse
(\code{rank = 1} for any vector).
}
}
\description{
Compute the Moore-Penrose's pseudo inverse of numeric vectors and
matrices (possibly non-square).
}
\details{
When argument \code{eps} is \code{NULL} (the default), it is set to
\code{.Machine$double.eps * max(NROW(X), NCOL(X)) * max(d)} where \code{d} is
the vector of singular values of \code{X}.

\code{inverse.matrix} internally calls \code{inverse.vector} when \code{X} is
a vector (including when it is a matrix with only one row or one column).

\code{inverse.vector} always turns its first argument into a vector.
}
\note{
The computations uses singular value decomposition as described in Section
\code{3.6.3} in \insertCite{petersen2012matrix;textual}{bigpcor} (see also
\insertCite{barnett1990matrices;textual}{bigpcor}). For rectangular matrices
(with two or more rows and columns), singular value decomposition is
performed using \link{svd}.
}
\examples{
#* Inverse of the vector [1, 2, 3]
inverse (1:3)

#* Inverse of the row vector [1, 2, 3]
inverse (t(1:3))

#* Inverse of the column vector [1, 2, 3]
inverse (t(t(1:3)))

#* Inverse of a rectangular matrix and its transpose
A <- matrix(c(5, 2, 3, 2, 3, 5), nrow = 3, ncol = 2)
A
inverse (A)

inverse (t(A))

#* Because A is tall and of full column rank, inverse (A) times A is identity
inverse (A) \%*\% A

#* Define B as:
B <- cbind(A, A[,1])
B

#* Because B is not of full rank, inverse B times B is not identity
inverse (B) \%*\% B
}
\references{
\insertAllCited{}
}
\seealso{
\link[bigpcor]{inverse} for the generic function.
}
