% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.R
\name{inverse}
\alias{inverse}
\title{Generalized Inverse of Numeric Objects}
\usage{
inverse(object, ...)
}
\arguments{
\item{object}{a numeric matrix-like object of a class with a
\code{inverse} method. For the default method, \code{object} is a numeric
vector or a matrix.}

\item{...}{additional argument passed to or from other methods.

Currently, the default method uses \code{eps}: a positive tolerance
below which any singular value of a matrix is considered zero.}
}
\value{
An object of the same class as \code{object}.

For the default method
\link{inverse.matrix}, a numeric \code{'matrix'} if \code{object} is of class
\code{matrix}, and a numeric vector if \code{object} is of class \code{vector}.
The returned object also has attributes:
\itemize{
\item \code{m} the minimum of the number of rows (if any),  and column, if any,
of \code{object};
\item \code{rank} the matrix rank of \code{object} or its returned inverse
(\code{rank = 1} for any vector).
}
}
\description{
\code{inverse} is a generic function whose default methods compute the
Moore-Penrose's generalized inverse of numeric vectors and matrices.
}
\details{
Additional \code{inverse} methods can be defined for specific classes.

For the default method, when \code{object} is a numeric vector, or a matrix
of only one row or one column, any \code{NA} or \code{NaN} is ignored.
If \code{object} is a matrix with two rows or more and two columns or more,
any \code{NA} or \code{NaN} in \code{object} causes error.
}
\seealso{
\link[bigpcor]{inverse.matrix} for the default method.
}
