% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tengenes}
\alias{tengenes}
\alias{fiftygenes}
\alias{hundredgenes}
\title{Simulated Fake Genomic Network Data}
\format{
An object of class \code{"data.frame"}.
}
\source{
Simulated
}
\usage{
data(tengenes)

data(fiftygenes)

data(hundredgenes)
}
\description{
Datasets mimicking genomic networks.
This is aimed at showcasting the selection of confounders of gene expressions
using package \code{bigpcor}.
}
\details{
Four simulated datasets are available.
Each dataset has three sets of variables,
\itemize{
\item a set 'V' with \code{p} variables, representing bi-allelic genetic variants
\item a set 'E' with \code{p} variables, representing gene expression values
\item a set 'C' with \code{u} variables, representing candidate confounders
for the expression values in 'E'.
}

For \code{tengenes}, \code{p=10} and \code{u = 20} with sample size \eqn{n=50}.

For \code{fiftygenes}, \code{p=50} and \code{u = 100} with sample size \eqn{n=300}.

For \code{hundredgenes}, \code{p=100} and \code{u = 200} with sample size \eqn{n=500}.
}
\examples{
data(tengenes)
\donttest{
pairs(tengenes[,11:20])
}
}
\keyword{datasets}
