% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.inner.inv.R
\name{update.inner.inv}
\alias{update.inner.inv}
\title{Rank-1 Update of Inverses of Inner Products}
\usage{
\method{update}{inner.inv}(object, add = TRUE, position = NULL, v, X, MP = NULL, ...)
}
\arguments{
\item{object}{numeric matrix \eqn{A} of size \eqn{p \times p}, inverse of
the inner product matrix \eqn{t(X) \times X}. \code{object} can have an
attribute 'X', which is required if \code{add = TRUE} and argument \eqn{X} is missing.}

\item{add}{logical, adding a new column into \eqn{X}?}

\item{position}{a positive integer, indicating the position \eqn{j} of:
\itemize{
\item{the vector \eqn{v} being added into \eqn{X} when \code{add = TRUE}}{, here \eqn{1 \leq j \leq p + 1} with default value \code{position}=\eqn{j}=\code{p+1};}
\item{the vector \eqn{v} being dropped from \eqn{X} when \code{add = FALSE}}{, here \eqn{1 \leq j \leq p} with default value \code{position}=\eqn{j}=\code{p}.}
}}

\item{v}{a length \eqn{n} numeric vector, required if \code{add = TRUE} (the
default) and not required nor used when \code{add = FALSE}.}

\item{X}{\eqn{n \times p} numeric matrix. Note that if \eqn{X} is the centered
version of another data matrix \eqn{Y} divided by the square root of the
sample degrees of freedom \eqn{n-1}, then \eqn{A} is the inverse of the
sample estimate of the covariance matrix of \eqn{Y}.

\eqn{X} can be missing as an argument, it must then be given as a named
attribute of \code{object} if \code{add = TRUE}. \eqn{X} is not required nor
used when \code{add = FALSE}.}

\item{MP}{logical or \code{NULL}, is \code{object} a Moore-Penrose inverse?
The default (\code{NULL}) looks for a \code{rank} attribute of \code{object}
and is \code{MP = TRUE} if this rank in less than the size of \code{object},
and \code{MP = FALSE} otherwise. If \code{object} has no \code{rank}
attribute, the rank of \code{object} is determined using QR decomposition
(\link{qr}).}

\item{...}{additional argument passed to or from other methods. Currently not
used.}
}
\value{
an object of class \code{"inner.inv"}, i.e. a matrix with a named
attribute \code{"X"} (which can be \code{NULL}, i.e. missing).
}
\description{
Update a square matrix \eqn{A}, inverse of the inner product matrix
\code{t(X) \%*\% X}, after \eqn{X} has been updated by including an
additional column \eqn{v} in \eqn{X}, or by removing one column from \eqn{X}.
}
\note{
The formula for a square non-singular matrix are taken from
\insertCite{khan2008updating;textual}{bigpcor} (see also Section \code{3.2.6}
in \insertCite{petersen2012matrix;textual}{bigpcor}).

The equivalent formula for pseudo-inverse was obtained by combining Eq. (168)
(Section \code{3.2.7}) in \insertCite{petersen2012matrix;textual}{bigpcor}
with Eq. (2.1) in \insertCite{baksalary2007particular;textual}{bigpcor}.
}
\references{
\insertAllCited{}
}
\seealso{
\link[bigpcor]{inverse} for computing pseudo-inverse of vector
and (non-square) matrices.
}
